use relative file path in CommandLineFormatter
This changes the format of the file name to add a leading "." so that
the path is a relative path instead of an absolute path.

Modern terminal programs allow control-clicking on paths to jump to a
file, so if this is a relative path, we can jump to the file. If it is
printed as an absolute path, this doesn't work.