Getting csharpier.msbuild happy with the new compilation errors being treated as errors by csharpier.
There didn't appear to be a good way to get msbuild happy with the different possible outcomes without adding the flag to treat compilation errors as warnings.
I tried using IgnoreStandardErrorWarningFormat and IgnoreExitCode in different combinations based on if CSharpier_Check was true or false but couldn't get every scenario working.