Modify VSCode extension to use dotnet tool list + only warn to instal… (#1323)
This modifies the VSCode extension in the following ways
- Instead of looking for dotnet-tools.json files in the current
directory and parent directories the extension runs `dotnet tool list`
and parses the outpud looking for the csharpier version.
- Instead of running `dotnet csharpier --version` to look for the global
version it repeats the above step with `dotnet tool list -g`
- The "You need to install csharpier" window was popping up every time
you opened a file if it couldn't find an install. The other extensions
were limited to showing the warning a single time. VSCode now only shows
it a single time.
- Files under `/DecompilationMetadataAsSourceFileProvider/` will never
have the warning show. This is where the c# dev kit decompiles code.

#1267