Set working directory for dotnet tool install in extensions (#1441)
This fixes #1439.

I've been able to verify for Rider and VS Code that it now correctly
uses our internal NuGet feed for the tool installation when a
`NuGet.config` is present in the repository. Couldn't test Visual Studio
because I don't have that, but considering that the code follows the
exact same structure for all extensions, I'm pretty confident it behaves
the same. Props to you for maintaining consistency between the
extensions by the way, it makes working on that code a blast!

---------

Co-authored-by: Bela VanderVoort <twobitbela@gmail.com>
Co-authored-by: Bela VanderVoort <bela.vandervoort@optimizely.com>