Undo the change to tool command (#1461)
If a tool has a command of `csharpier` then to invoke it as a local tool
you need to use `dotnet csharpier`. Only a globally installed tool works
with `csharpier`. It doesn't seem worth the trouble to change the
command because of that.

If https://github.com/dotnet/sdk/issues/14626 is done, this can be
revisited.

closes #1450