Set server pwd to cache dir
The ASP.NET application watches configures the 
content root to be the process working directory by default.
To watch for changes to its configuration (appsettings.json)
it also installs a recursive file system watch.
This is highly problematic when the working directory is the
repository root because then, csharpier will ask to be notified
of changes in the entire repository (including .git, .idea,
node_modules, bin, obj etc.).