feat: reuse collection types for `DocFitter` (#1491)
Instead of creating a new instance of `StringBuilder` and
`Stack<PrintCommand>` every time `DocFitter.Fit` is called we instead
create a single instance of each in `DocPrinter` and pass it into
`DocFitter.Fit`.

This could be replaced by using `ValueStringBuilder` and
`ValueListBuilder`, the latter would be beneficial as unlike `Stack<T>`
we can access items by index instead of the `O(n)` lookup time with
`ElementAt`.

### Before
| Method | Mean | Error | StdDev | Median | Gen0 | Gen1 | Gen2 |
Allocated |
|----------------------
|---------:|--------:|--------:|---------:|-----------:|----------:|----------:|----------:|
| Default_CodeFormatter | 231.5 ms | 4.62 ms | 8.78 ms | 228.0 ms |
11000.0000 | 4000.0000 | 1000.0000 | 95.82 MB |

### After
| Method | Mean | Error | StdDev | Gen0 | Gen1 | Gen2 | Allocated |
|----------------------
|---------:|--------:|--------:|-----------:|----------:|----------:|----------:|
| Default_CodeFormatter | 227.0 ms | 4.51 ms | 6.33 ms | 10000.0000 |
4000.0000 | 1000.0000 | 90.03 MB |

Co-authored-by: Bela VanderVoort <twobitbela@gmail.com>