Modifying CSharpier to support reading gitignore files to determine what files are ignored. (#1484)
Closes #631 

CSharpier now works as follows.

- .gitignore files are considered when determining if a file will be
ignored. A .gitignore file at the same level as a given file will take
priority over a .gitignore file above it in the directory tree.
- If a .csharpierignore file is present at the same level or anywhere
above the given file in the tree and it contains a pattern for a given
file, that will take priority.
- The patterns within .csharpierignore work the same as a .gitignore,
with patterns lower in the file taking priority over patterns above
- CSharpier does not currently look further up the directory tree for
additional .csharpierignore files if it finds one. But it does look for
.gitignore files. If there is demand this could be added later.