Use alternative library to parse csharpier- and git-ignore files (#1591)
Using gitignore parser library instead of building ignore rules
ourselves improves csharpier performance drastically even with large
repositories and a lot of ignore rules.

This would introduce a potentially breaking change as the rules are now
case-sensitive by default since git is as well.

I had to remove one unit test as we would not parse the ignore-files
ourselves anymore as well as alter one which was testing
case-insensitive paths.