fix: Null coalescing not own its line with complex left-hand side (#1770)
There was an issue where if the left-hand side of `??` was even slightly
complex, the right-hand side would become part of the group and could
not have its own line.

I reconsidered the conditions for grouping and changed it from judging
the complexity of the left-hand side by a threshold to a condition where
"the left-hand side is an InvocationExpression and consists solely of
MemberAccessExpression that does not include an InvocationExpression".

fix #1769